requirements.txt
Flask
Werkzeug
google-cloud-storage
google-generativeai